move router to the ui module prepping for Pane adjustment
