+++ src/ui/mod.rs
@@ -3,3 +3,4 @@ pub use manager::UiManager;
 mod components;
 mod manager;
 mod panes;
+mod router;
